Fix setup of TestSuite

+++ TemportalEngine/source/Window.cpp
@@ -21,7 +21,7 @@ Window::Window(uSize width, uSize height, char const * title)
 	, mpTitle(title)
 {
 	this->mIsPendingClose = false;
-	this->mpHandle = SDL_CreateWindow(mpTitle, SDL_WINDOWPOS_CENTERED, SDL_WINDOWPOS_CENTERED, mWidth, mHeight, 0);
+	this->mpHandle = SDL_CreateWindow(mpTitle, SDL_WINDOWPOS_CENTERED, SDL_WINDOWPOS_CENTERED, (int)mWidth, (int)mHeight, 0);
 	if (!this->isValid())
 	{
 		DeclareLog("Window").log(logging::ECategory::ERROR,
